backend/requirements.txt
# Sport Dashboard - Backend Dependencies
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
pydantic
pytz
psycopg2-binary
alembic
passlib
bcrypt
python-jose
cryptography
python-multipart
email-validator
